move alertavailablefor out of attributes maker because it is also needed for mma-membership
